cleanup: move formatting into utils folder
