+++ clients/trieve-shopify-extension/utils/formatting.ts
